schema: Change Epoch view into a table
As a view, this was costly to update and is better implemented as a
table.

This table is only used by the Epoch plugin, but managing tables
required by the plugin system outside the migration framework in the
explorer-db-node is incredibly difficult.

For now, add the Epoch table as part of the normal migrations but
it will not be used by the normal explorer-db-node.