node: Use Async.race_ to ensure proper behavior
Previously, exceptions thrown within 'runDbThread' would be stored
in a 'TVar' to be read from 'Async.wait'. Unfortunately 'wait' was
never called so the exception would never get reported.

After some help with the network protocol types I was able to switch
to using 'race_' which has sane behaviour with regard to exceptions.