Minor type clarification: use Tip type
The plugin API for inserting a block takes a block and the server's
current tip block number. But since it uses the types ByronBlock and
BlockNo, it's not immediately clear that this is the BlockNo of the tip
rather than a plausible first guess which is that it's the block number
of the block being inserted.

To avoid this danger of confusion, switch the type from BlockNo to the
type Tip ByronBlock. Then the type makes clear that it is the tip and
not about the block itself. It is of course still possible to get the
block number from the tip.